requirements.txt

fastapi
uvicorn
redis
mysql-connector-python
python-dotenv
requests
elevenlabs
groq
dateparser


google-api-python-client 
google-auth 
google-auth-oauthlib